Key the live cadence on the read, and move the moment into the kit
CI's fe-file-length ratchet caught the first cut: it grew six files that
are already over the 500-line cap, and that list may only move shorter.
The fix is the one the gate was asking for.

The live cadence moves off the four record hooks and onto the client,
keyed on what a read IS. `isRecordRead` derives its corpus from
TIMELINE_SEED_KEYS — the table that already knows which key carries a
record — so a record kind joins by existing rather than by a screen
remembering to spread options in. Per-hook options fail the way lists
fail: the fifth record page is written without them, nothing says so,
and it serves a stale answer that looks exactly like a fresh one. The
deal's model-written briefing stays out because its key is the status
card's, and the suite pins that exclusion.

The moment's vocabulary moves from the contact page into the record360
kit, where every type in it (Grounding, StandingTone) already lived. The
account brief had been importing MOMENT_RULE_LABEL, standingTone and
momentGrounding across from a screen — the borrowing the kit exists to
end — and MomentRow goes with them. persontoday.tsx drops under the
500-line cap for good, so its waiver entry is removed and the cap
re-arms on it.

The contact-page case moves out of personpage.test.tsx, which is at its
ceiling, into a persontoday suite that mounts the card directly. The
live-record suite now runs against createQueryClient() rather than a
bare client, so it proves the policy that ships.

Also caught locally: the docs page-length gate. The two new sections in
company-record-page.md are one, trimmed to the page's budget.
