+++ scripts/fe-file-length-waivers.txt
@@ -113,7 +113,6 @@
 1208 frontend/src/screens/personprovider.test.tsx
 941 frontend/src/screens/personprovider.tsx
 1575 frontend/src/screens/personrail.tsx
-531 frontend/src/screens/persontoday.tsx
 1200 frontend/src/screens/privacy.test.tsx
 1245 frontend/src/screens/privacy.tsx
 571 frontend/src/screens/project360.tsx
